// src/interfaces/queryResults/antsFeedData.ts
export default interface AntsFeedData {
  vrm_trm?: string | undefined;
  make?: string | undefined;
  model?: string | undefined;
  wheelplan?: string | undefined;
  test_date?: string | undefined;
  weight_before_test?: bigint | undefined;
  weight_after_test?: bigint | undefined;
  DOE_reference?: string | undefined;
  tech_record_date?: string | undefined;
}